[WIP architecture] : split lazy-modules
