+++ src/app/app.module.ts
@@ -4,9 +4,10 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { WebfullstackDesignSystemModule } from "webfullstack-design-system";
+import { HeaderComponent } from './components/header/header.component';
 
 @NgModule({
-  declarations: [AppComponent],
+  declarations: [AppComponent, HeaderComponent],
   imports: [BrowserModule, AppRoutingModule, WebfullstackDesignSystemModule],
   providers: [],
   bootstrap: [AppComponent],
