+++ src/app/views/crypto/components/crypto-card/crypto-card.component.html
@@ -1,0 +1,18 @@
+<lib-card class="crypto-card">
+    <flex-div
+      *ngIf="crypto"
+      justifyContent="space-between"
+      class="crypto-card__content"
+      [fullWidth]="true"
+    >
+      <flex-div alignItems="center">
+        <img [src]="crypto.img" alt="bitcoin" class="crypto-card__img" />
+        <flex-div direction="column">
+          <p>{{ crypto.name }}</p>
+          <p>{{ crypto.abbreviation }}</p>
+        </flex-div>
+      </flex-div>
+      <p>{{ crypto.eur_price }}€</p>
+    </flex-div>
+  </lib-card>
+
