+++ src/app/views/crypto/components/crypto-card/crypto-card.component.scss
