+++ src/app/views/crypto/models/crypto.ts
